tests/wpt/web-platform-tests/webvtt/parsing/file-parsing/tests/settings-position.html

<!doctype html>
<!-- DO NOT EDIT! This file and support/settings-position.vtt are generated. -->
<!-- See /webvtt/parsing/file-parsing/README.md -->
<meta charset=utf-8>
<title>WebVTT parser test: settings, position</title>
<link rel="help" href="https://w3c.github.io/webvtt/#collect-webvtt-cue-timings-and-settings">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var t = async_test('settings, position');
t.step(function(){
    var video = document.createElement('video');
    var track = document.createElement('track');
    assert_true('src' in track, 'track element not supported');
    track.src = 'support/settings-position.vtt';
    track['default'] = true;
    track.kind = 'subtitles';
    track.onload = this.step_func(trackLoaded);
    track.onerror = this.step_func(trackError);
    video.appendChild(track);
    document.body.appendChild(video);
});

function trackLoaded(event) {
    var track = event.target;
    var video = track.parentNode;
    var cues = video.textTracks[0].cues;
    {
assert_equals(cues.length, 22);

var valid_positions = [
    [1, 'auto'],
    [100, 'auto'],
    [1, 'auto'],
    [1.5, 'auto'],
    [1, 'line-left'],
    [1, 'center'],
    [1, 'line-right'],
    [1, 'auto'],
];
valid_positions.forEach(function(pair, index) {
    var position = pair[0];
    var positionAlign = pair[1];

    assert_equals(cues[index].position, position, 'Failed with cue ' + index);
    assert_equals(cues[index].positionAlign, positionAlign, 'Failed with cue ' + index);
});

for (var i = 0; i < 14; i++) {
    var index = valid_positions.length + i;

    assert_equals(cues[index].position, 'auto', 'Failed with cue ' + index);
    assert_equals(cues[index].positionAlign, 'auto', 'Failed with cue ' + index);
}

    }
    this.done();
}

function trackError(e) {
    assert_unreached('got unexpected error event');
}
</script>
