third_party/blink/web_tests/external/wpt/webvtt/parsing/file-parsing/tests/settings-position-expected.txt

This is a testharness.js-based test.
[FAIL] settings, position
  assert_equals: Failed with cue 0 expected (string) "auto" but got (undefined) undefined
Harness: the test ran to completion.
